Fix formatting issues, separate AggregateParams
