+++ pipeline_dp/__init__.py
@@ -1,7 +1,7 @@
 from pipeline_dp.budget_accounting import BudgetAccountant
-from pipeline_dp.dp_engine import AggregateParams
+from pipeline_dp.aggregate_params import AggregateParams
+from pipeline_dp.aggregate_params import Metrics
 from pipeline_dp.dp_engine import DataExtractors
-from pipeline_dp.dp_engine import Metrics
 from pipeline_dp.dp_engine import DPEngine
 from pipeline_dp.pipeline_operations import BeamOperations
 from pipeline_dp.pipeline_operations import SparkRDDOperations
