+++ pipeline_dp/dp_engine.py
@@ -1,52 +1,12 @@
 """DP aggregations."""
 
-from enum import Enum
-from typing import Any, Callable, Iterable
+from typing import Callable
 
 from dataclasses import dataclass
+from pipeline_dp.aggregate_params import AggregateParams
 from pipeline_dp.budget_accounting import BudgetAccountant
 from pipeline_dp.pipeline_operations import PipelineOperations
 from pipeline_dp.report_generator import ReportGenerator
-
-
-class NoiseKind(Enum):
-  LAPLACE = 'laplace'
-  GAUSSIAN = 'gaussian'
-
-
-class Metrics(Enum):
-  COUNT = 'count',
-  PRIVACY_ID_COUNT = 'privacy_id_count',
-  SUM = 'sum',
-  MEAN = 'mean',
-  VAR = 'variance'
-
-
-@dataclass
-class AggregateParams:
-  """Specifies parameters for function DPEngine.aggregate()
-
-  Args:
-    metrics: Metrics to compute.
-    max_partitions_contributed: Bounds the number of partitions in which one
-      unit of privacy (e.g., a user) can participate.
-    max_contributions_per_partition: Bounds the number of times one unit of
-      privacy (e.g. a user) can contribute to a partition.
-    low: Lower bound on a value contributed by a unit of privacy in a partition.
-    high: Upper bound on a value contributed by a unit of privacy in a
-      partition.
-    public_partitions: a collection of partition keys that will be present in
-      the result.
-  """
-
-  metrics: Iterable[Metrics]
-  max_partitions_contributed: int
-  max_contributions_per_partition: int
-  low: float = None
-  high: float = None
-  budget_weight: float = 1
-  public_partitions: Any = None
-
 
 @dataclass
 class DataExtractors:
